requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
idna==3.11
numpy==2.4.4
packaging==26.0
pandas==3.0.2
proto-plus==1.27.2
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
